# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: string;
  title: string;
  description?: string | null;
  dueDate?: Date | null;
  status: TaskStatus;
  createdAt: Date;
}