# lib/api-zod/src/generated/types/updateTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskStatus } from "./updateTaskStatus";

export interface UpdateTask {
  title?: string;
  description?: string | null;
  dueDate?: Date | null;
  status?: UpdateTaskStatus;
}